Authenticate yourself with BrandsEye

Provides a means to authenticate yourself with the BrandsEye API. This is done using an API key that your account manager representative has provided for you.

authenticate(key, save = FALSE)

Arguments

key
Your API key. If you don't know this, your account manager can give you one.

save
Set to true if you would like this key to be saved to disc and loaded automatically the next time that you use this library.

Value

Your authentication details

Details

All authentication with the BrandsEye API is done via a special, user specific key. This key can be provided to you by your account manager.

It's possible to also have your key saved to disc and automatically loaded for you when you next use the library. Set the save parameter to TRUE in order to do this.

If you would like to see who you are currently logged in as, see the whoami() function.

See also

Examples

if (FALSE) {

authenticate("adfd42345f534fgdfgd")

}