Read account logs

Returns a tibble of logs for the account, in the last period stated.

logs(x, from, .show.progress)

# S3 method for character
logs(x, from = (Sys.Date() %m-% months(1)), .show.progress = interactive())

# S3 method for brandseyer2.account
logs(x, from = (Sys.Date() %m-% months(1)), .show.progress = interactive())

# S3 method for list
logs(x, from = (Sys.Date() %m-% months(1)), .show.progress = interactive())

# S3 method for data.frame
logs(x, from = (Sys.Date() %m-% months(1)), .show.progress = interactive())

Arguments

x
An object to read logs from.

from
A date indicating from when to read dates.

.show.progress
A logical indicating whether to show a progress bar or not. By default, this only shows a progress bar in interactive environments. Also, progress bars will only be shown if 3 or more accounts are being examined.

Value

A tibble of log information, including the user that perform the action, and how many times the action was performed.

Details

Logs can be read from various different kinds of object. An account() object is the most basic, although logs can also be read from a list of account objects (in which case the returned tibble will include an account code column), or from a tibble containing an account code column, such as that returned by account_list().

Methods (by class)

character: Reach for an account code character vector.

brandseyer2.account: For account() objects.

list: For a list of accounts

data.frame: Read for a data frame, such as returned by account_list()

Useful logging functions

logs_retrosent() filters log information to find retrosend specific information.

Examples

if (FALSE) {
# Defaults to the last month of data
logs("TEST01AA")

# The last week
logs("TEST01AA", Sys.Date() - lubridate::weeks(1))

# The last day
logs("TEST01AA", Sys.Date() - lubridate::days(1))

# The last three months
library(lubridate)
logs("TEST01AA", Sys.Date() %m-% months(1))
}

if (FALSE) {
# Read logs for an account
account("TEST01AA") %>%
  logs()
}

if (FALSE) {
# Read for a list of accounts
account("TEST01AA", "TEST02AA") %>%
  logs()
}
if (FALSE) {
account_list() %>%
  logs()
}