Fetch topics for an account or tibble of mentions

Fetches the topics that are used in an account, or those that appear on a tibble of mentions.

# S3 method for brandseyer2.account
topics(x, ...)

# S3 method for list
topics(x, ...)

# S3 method for data.frame
topics(x, ..., ac = attr(x, "account"), na.rm = TRUE)

topics(x, ...)

Arguments

x
Objects to return topic data from.

...
Additional arguments for other methods.

ac
An optional account object from which to take topic information.

na.rm
Whether to remove mentions that have no topics.

Value

A tibble of topic information. Includes the name and topic description.

Methods (by class)

brandseyer2.account: Get topics for a single account.

list: Get topics for a list of accounts

data.frame: Fetch topic information from a tibble of mention data.

Examples

account("TEST01AA") %>%
  topics()

#> # A tibble: 3 x 7
#>      id name          namespace description       deleted children  is_parent
#>   <dbl> <chr>         <chr>     <chr>             <lgl>   <list>    <lgl>    
#> 1    10 Parent topic  topic     Has children      FALSE   <dbl [2]> TRUE     
#> 2    11 Child topic 1 topic     I am a child      FALSE   <NULL>    FALSE    
#> 3    12 Child topic 2 topic     I am also a child FALSE   <NULL>    FALSE    

accounts(c("TEST01AA", "TEST02AA")) %>%
  topics()

#> # A tibble: 6 x 8
#>   account     id name          namespace description deleted children is_parent
#>   <chr>    <dbl> <chr>         <chr>     <chr>       <lgl>   <list>   <lgl>    
#> 1 TEST01AA    10 Parent topic  topic     Has childr… FALSE   <dbl [2… TRUE     
#> 2 TEST01AA    11 Child topic 1 topic     I am a chi… FALSE   <NULL>   FALSE    
#> 3 TEST01AA    12 Child topic 2 topic     I am also … FALSE   <NULL>   FALSE    
#> 4 TEST02AA  1010 Parent topic  topic     Has childr… FALSE   <dbl [2… TRUE     
#> 5 TEST02AA  1011 Child topic 1 topic     I am a chi… FALSE   <NULL>   FALSE    
#> 6 TEST02AA  1012 Child topic 2 topic     I am also … FALSE   <NULL>   FALSE