Determine the parent of tags given a particular tree

Adds a column called parent to a table. This is the ID of a tag's parent tag. Since tags can have multiple parents, we need to know what the ultimate root parent to determine "parenthood" is. This is done using the parent_id parameter.

with_tag_parents(tags, parent_id)

Arguments

tags
A tibble of tags to find parents for

parent_id
A root parent that all children are determined from

Value

The same tags tibble, but annotated with a parent column

Examples

# Add parents to tags
account("TEST01AA") %>%
  tags() %>%
  with_tag_parents(10)

#> Source: local data frame [6 x 8]
#> Groups: <by row>
#> 
#> # A tibble: 6 x 8
#>      id name       namespace  description     deleted children is_parent parent
#>   <int> <chr>      <chr>      <chr>           <lgl>   <list>   <lgl>      <dbl>
#> 1     1 tag1       tag        <NA>            FALSE   <NULL>   FALSE         NA
#> 2     2 tag2       tag        Deleted         TRUE    <NULL>   FALSE         NA
#> 3    10 Parent to… topic      Has children    FALSE   <int [2… TRUE          NA
#> 4    11 Child top… topic      I am a child    FALSE   <NULL>   FALSE         10
#> 5    12 Child top… topic      I am also a ch… FALSE   <NULL>   FALSE         10
#> 6  1001 Tree of T… topic_tree Topics come fr… FALSE   <int [1… TRUE          NA

# See what root brands we have
account("TEST01AA") %>%
  topic_trees()

#> # A tibble: 1 x 5
#>      id name           description           deleted children 
#>   <int> <chr>          <chr>                 <lgl>   <list>   
#> 1  1001 Tree of Topics Topics come from here FALSE   <int [1]>

# Tag 1001 is a topic tree. Let's find
# topics, with parents, that are part of this tree.
account("TEST01AA") %>%
  tags() %>%
  with_tag_parents(1001) %>%
  dplyr::filter(namespace == 'topic') %>%
  dplyr::arrange(desc(is_parent), name)

#> # A tibble: 3 x 8
#>      id name        namespace description     deleted children is_parent parent
#>   <int> <chr>       <chr>     <chr>           <lgl>   <list>   <lgl>      <dbl>
#> 1    10 Parent top… topic     Has children    FALSE   <int [2… TRUE        1001
#> 2    11 Child topi… topic     I am a child    FALSE   <NULL>   FALSE         10
#> 3    12 Child topi… topic     I am also a ch… FALSE   <NULL>   FALSE         10