List the dashboards in an account

In Analyse, BrandsEye's analysis tool, people can set up dashboards to query our data. This function exposese those dashboards as a tibble, allowing you to query sections, metrics and their associated filters.

dashboards(x, d, ...)

# S3 method for brandseyer2.account
dashboards(x, d, section, ...)

Arguments

x
An account object to query.

d
An optional value to filter the dashboards on. Can be a vector of integer IDs for the wanted dashboards, or a vector of characters giving words to perform partial string matches on dashboard names with.

...
Additional parameters for methods

section
An optional numeric vector of section IDs, or character vector to match against section titles. This specifies which sections to return, similarly to calling sections(). You can use the special string '.' to select all sections.

Value

A tibble of dashboard data.

Details

Dashboards can be filtered in various ways. The second argument, d, can be either an integer or list of integers, specifying the particular dashboard IDs that are wanted. The second argument can also be a character vector, to give partial matches against dashboard names. Case is ignored.

Methods (by class)

brandseyer2.account: Read dashboard data for an account() object.

See also

Examples

# NOT RUN {
# Get all dashboards.
account("TEST01AA") %>%
  dashboards()

# Get the dashboard with ID 1.
account("TEST01AA") %>%
  dashboards(1)

# Get dashboards 1 and 21.
account("TEST01AA") %>%
  dashboards(c(1, 21))

# Get dashboards whose name contains the string 'overview'.
account("TEST01AA") %>%
  dashboards("overview")

# Get dashboards whose name contains the string 'overview' or the string 'who'.
account("TEST01AA") %>%
  dashboards(c("overview", "who"))

# }