Finding retrosends for accounts

logs_retrosent() can be used to find retrosends that have occurred for a particular account() or list of accounts. If you want to specify a date range, use logs() and pipe its results through to logs_retrosent().

logs_retrosent(x, ...)

# S3 method for data.frame
logs_retrosent(x, ...)

# S3 method for character
logs_retrosent(x, ..., .show.progress = interactive())

# S3 method for brandseyer2.account
logs_retrosent(x, ...)

# S3 method for list
logs_retrosent(x, ..., .show.progress = interactive())

Arguments

x
A data frame of logs, or an account object (or list of account objects)

...
Extra arguments

.show.progress
When reading logs, whether to show a progress bar or not. See logs().

Value

A tibble giving retrosend information, such as the crowd, priority and filter used. The requested column is the number of mentions requested to be sent to the crowd, while the sent column is the number actually sent.

Methods (by class)

data.frame: Reading from a data.frame or tibble of log information, such as that given by logs().

character: From a vector of account codes.

brandseyer2.account: From an account object

list: From a list of account objects.

Examples

if (FALSE) {

account("TEST01AA") %>%
  logs_retrosent()

account_list() %>%
  logs() %>%
  logs_retrosent()

# Pull date for a particular time range
account("TEST01AA") %>%
  logs(Sys.Date() - lubridate::weeks(1)) %>%
  logs_retrosent()

# Pull data from a particular date up to today.
account("TEST01AA") %>%
  logs(lubridate::ymd("2018/08/01")) %>%
  logs_retrosent()
}