Fetch phrases for an accounts and mentions

Fetches phrase information, returned as a tibble, for the given account. All of this information is included in brands()

# S3 method for brandseyer2.account
phrases(x, ...)

# S3 method for list
phrases(x, ...)

# S3 method for data.frame
phrases(x, ..., ac = attr(x, "account"))

phrases(x, ...)

Arguments

x
An object to read phrases from, such as an account() object, or a tibble of mentions.

...
Additional arguments for other methods.

ac
An optional account object from which to take phrase information.

Value

A tibble of phrase information.

Methods (by class)

brandseyer2.account: Read phrases for only a single account

list: Create a table of phrases for the list of accounts given

data.frame: Fetch phrase information from a tibble of mention data.

Note

mentions() should be called with select = "phrases" to ensure that phrase information is returned with the mentions.

Examples

# Fetch phrases for a single account
account("TEST01AA") %>%
  phrases()

#> # A tibble: 9 x 5
#>   phrase.id brand.id query      inactive deleted
#>       <dbl>    <dbl> <chr>      <lgl>    <lgl>  
#> 1       102        5 branded    FALSE    FALSE  
#> 2       102        6 branded    FALSE    FALSE  
#> 3       101        5 brander    FALSE    FALSE  
#> 4       101        6 brander    FALSE    FALSE  
#> 5         5        3 niece      FALSE    FALSE  
#> 6         3        2 sibling    FALSE    FALSE  
#> 7         2        1 your brand FALSE    FALSE  
#> 8         4        2 oh wow     TRUE     FALSE  
#> 9         1        1 brand      FALSE    TRUE   

accounts(c("TEST01AA", "TEST02AA")) %>%
  phrases()

#> # A tibble: 11 x 6
#>    account  phrase.id brand.id query      inactive deleted
#>    <chr>        <dbl>    <dbl> <chr>      <lgl>    <lgl>  
#>  1 TEST01AA       102        5 branded    FALSE    FALSE  
#>  2 TEST01AA       102        6 branded    FALSE    FALSE  
#>  3 TEST01AA       101        5 brander    FALSE    FALSE  
#>  4 TEST01AA       101        6 brander    FALSE    FALSE  
#>  5 TEST01AA         5        3 niece      FALSE    FALSE  
#>  6 TEST01AA         3        2 sibling    FALSE    FALSE  
#>  7 TEST01AA         2        1 your brand FALSE    FALSE  
#>  8 TEST01AA         4        2 oh wow     TRUE     FALSE  
#>  9 TEST01AA         1        1 brand      FALSE    TRUE   
#> 10 TEST02AA       102      100 branded    FALSE    FALSE  
#> 11 TEST02AA       101      100 brander    FALSE    FALSE