Find root brands in an account

Root brands are the brands around which we collect mentions, and determine sentiment. Sentiment is always towards the brand.

root_brands(x, includeDeleted, includeArchived, short)

# S3 method for data.frame
root_brands(x, includeDeleted = FALSE,
  includeArchived = FALSE, short = TRUE)

# S3 method for brandseyer2.account
root_brands(x, includeDeleted = FALSE,
  includeArchived = FALSE, short = TRUE)

# S3 method for list
root_brands(x, includeDeleted = FALSE,
  includeArchived = FALSE, short = TRUE)

Arguments

x
The object to find root brands for.

includeDeleted
Whether to include deleted root brands. By default, only shows non-deleted brands.

includeArchived
Whether to include archived root brands. By default, only shows non-archived brands.

short
Include only the most import bits of brand information, such as the name and ID.

Value

A tibble containing rows only for the root brands of an account. Otherwise, like the table returned in brands().

Methods (by class)

data.frame: Find root brands from a tibble or data.frame. The data.frame should have id, parent, and deleted columns, just as the tibble returned from brands() does.

brandseyer2.account: Returns root brand information from an account object.

list: Returns root brand information from a list of account objects.

See also

Examples

# Get root brands for an account
account("TEST01AA") %>%
  root_brands()

#> # A tibble: 2 x 4
#>      id name            deleted archived
#>   <int> <chr>           <lgl>      <dbl>
#> 1     1 Your Brand Here FALSE         NA
#> 2     2 Sibling Brand   FALSE         NA

# Get root brands from a tibble or data.frame
account(c("TEST01AA", "TEST02AA")) %>%
  brands(short = FALSE) %>%
  root_brands()

#> Joining, by = "id"

#> # A tibble: 3 x 5
#>   account     id name               deleted archived
#>   <chr>    <dbl> <chr>              <lgl>      <dbl>
#> 1 TEST01AA     2 Sibling Brand      FALSE         NA
#> 2 TEST02AA   100 The Best of Brands FALSE         NA
#> 3 TEST01AA     1 Your Brand Here    FALSE         NA

# Show fuller root brand information
account("TEST01AA") %>%
  root_brands(short = FALSE)

#> # A tibble: 2 x 12
#>      id parent name  tier  deleted schema filter topic_tree_id sentiment_rate
#>   <int>  <int> <chr> <chr> <lgl>   <chr>  <chr>          <dbl>          <dbl>
#> 1     2     NA Sibl… TOPI… FALSE   b0000… locat…           100            0.8
#> 2     1     NA Your… ESSE… FALSE   <NA>   <NA>              NA            0  
#> # ... with 3 more variables: topic_rate <dbl>, archived <dbl>, phrases <list>