Update data stored on a mention

This function lets you update data that is stored on a mention. For instance, you can add and rmeove tags from mentions.

update_mentions(x, filter, auto.confirm = FALSE, ...)

# S3 method for brandseyer2.account.v4
update_mentions(x, filter, auto.confirm,
  ..., relevancy, relevancyVerified, addTags, removeTags, sentiment,
  sentimentVerified, media, gender, language, race, addPhrases,
  removePhrases, addCrowdJobs, removeCrowdJobs, updateAuthor, location)

Arguments

x
The account object.

filter
Defines the mentions to update.

auto.confirm
Some updates, such as updating sentiment, can badly effect an account. This function will usually ask if you want to proceed. Setting this to TRUE will turn this behaviour off.

...
Arguments for other methods.

relevancy
A string indicating whether mentions are relevant or not.

relevancyVerified
A boolean indicating whether to consider the relevancy as verifed or not.

addTags
A list of tag IDs to add to the mention.

removeTags
A list of tag IDs to remove.

sentiment
An integer indicating sentiment.

sentimentVerified
A boolean indicating whether to consider sentiment as verified or not.

media
A string indicating the category of the mentions, such as CONSUMER or PRESS

gender
A string indicating the author's gender, such as MALE, FEMALE, OTHER, UNKNOWN.

language
A two letter, lowercase string, giving the iso 639-1 language code for the mention.

race
A string giving the race of the author

addPhrases
A list of integer IDs for phrases to add to the mention

removePhrases
A list of integer IDs for the phrases to remove from the mention

addCrowdJobs
A list of integer IDs representing crowd jobs related to this mention, to be added to the mention.

removeCrowdJobs
A list of integer IDs representing crowd jobs to be removed from this mention.

updateAuthor
Whether the stored author information related to this mention should be updated with any of the information related to the author on this mention, such as race, gender, language or media.

location
A string giving location information.

Value

Nothing

Methods (by class)

brandseyer2.account.v4:

Note

Setting sentiment and relevancy will affect crowd data. You should strongly think about setting relevancyVerified or sentimentVerified as appropriate, if you decide to edit sentiment or relevancy.